Moved data_types / data_traits to more suitable folder
