+++ include/sparrow/arrow_array_schema_proxy.hpp
@@ -20,7 +20,6 @@
 #include <type_traits>
 #include <utility>
 
-#include "sparrow/array/data_type.hpp"
 #include "sparrow/arrow_interface/arrow_array.hpp"
 #include "sparrow/arrow_interface/arrow_array/private_data.hpp"
 #include "sparrow/arrow_interface/arrow_flag_utils.hpp"
@@ -29,6 +28,7 @@
 #include "sparrow/buffer/buffer_view.hpp"
 #include "sparrow/buffer/dynamic_bitset.hpp"
 #include "sparrow/c_interface.hpp"
+#include "sparrow/types/data_type.hpp"
 #include "sparrow/utils/contracts.hpp"
 
 namespace sparrow
