+++ include/sparrow/layout/struct_layout/struct_value.hpp
@@ -15,8 +15,8 @@
 #pragma once
 
 #include "sparrow/config/config.hpp"
-#include "sparrow/array/data_traits.hpp"
 #include "sparrow/layout/array_base.hpp"
+#include "sparrow/types/data_traits.hpp"
 #include "sparrow/utils/memory.hpp"
 
 namespace sparrow
